@@ -83,8 +85,8 @@
       <plugin>
         <artifactId>maven-compiler-plugin</artifactId>
         <configuration>
-          <source>1.8</source>
-          <target>1.8</target>
+          <source>17</source>
+          <target>17</target>
           <encoding>${project.build.sourceEncoding}</encoding>
         </configuration>
       </plugin>
