@@ -1,7 +1,7 @@
 package pl.matsuo.core.model.numeration;
 
-import javax.persistence.Entity;
-import javax.validation.constraints.NotNull;
+import jakarta.persistence.Entity;
+import jakarta.validation.constraints.NotNull;
 import lombok.Getter;
 import lombok.Setter;
 import pl.matsuo.core.model.AbstractEntity;
@@ -16,6 +16,7 @@ public class NumerationSchema extends AbstractEntity {
   protected Integer maxValue;
   protected String code;
   protected String pattern;
+
   /** How not numeration instances should be created basing on data contained in this schema. */
   protected String creationStrategy;
 }
