@@ -1,19 +1,19 @@
 package pl.matsuo.core.model.print;
 
-import static javax.persistence.CascadeType.ALL;
-import static javax.persistence.InheritanceType.SINGLE_TABLE;
+import static jakarta.persistence.CascadeType.ALL;
+import static jakarta.persistence.InheritanceType.SINGLE_TABLE;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
+import jakarta.persistence.Entity;
+import jakarta.persistence.Inheritance;
+import jakarta.persistence.OneToMany;
+import jakarta.persistence.OrderColumn;
+import jakarta.persistence.Transient;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.function.Function;
 import java.util.function.Supplier;
-import javax.persistence.Entity;
-import javax.persistence.Inheritance;
-import javax.persistence.OneToMany;
-import javax.persistence.OrderColumn;
-import javax.persistence.Transient;
-import javax.validation.constraints.NotNull;
+import jakarta.validation.constraints.NotNull;
 import lombok.Getter;
 import lombok.Setter;
 import pl.matsuo.core.model.kv.KeyValueEntity;
@@ -28,6 +28,7 @@ import pl.matsuo.core.service.facade.IFacadeAware;
 public class KeyValuePrint extends KeyValueEntity implements IFacadeAware {
 
   @NotNull private Class<? extends IPrintFacade> printClass;
+
   /** Reference to entity on which print is based. */
   private Long idEntity;
 
