@@ -1,6 +1,6 @@
 package pl.matsuo.core.model.print;
 
-import javax.persistence.Entity;
+import jakarta.persistence.Entity;
 import pl.matsuo.core.model.kv.KeyValueEntity;
 
 @Entity
