@@ -7,7 +7,7 @@ import static java.lang.annotation.RetentionPolicy.RUNTIME;
 import java.lang.annotation.Documented;
 import java.lang.annotation.Retention;
 import java.lang.annotation.Target;
-import javax.validation.Constraint;
+import jakarta.validation.Constraint;
 import pl.matsuo.core.model.AbstractEntity;
 
 @Constraint(validatedBy = {EntityReferenceValidator.class})
