@@ -12,7 +12,7 @@ import static pl.matsuo.core.util.SecurityUtil.passwordHash;
 import java.util.Date;
 import java.util.List;
 import java.util.function.Function;
-import javax.validation.Valid;
+import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpEntity;
