@@ -3,13 +3,16 @@ package pl.matsuo.core.test;
 import static java.util.Optional.ofNullable;
 import static pl.matsuo.core.util.function.FunctionalUtil.runtimeEx;
 
-import java.io.ByteArrayInputStream;
+import lombok.Getter;
+import org.apache.pdfbox.Loader;
 import org.apache.pdfbox.pdmodel.PDDocument;
 import org.apache.pdfbox.text.PDFTextStripper;
 
 public class CreatePdfResult {
 
+  @Getter
   private String html;
+  @Getter
   private byte[] pdf;
   private String pdfString;
 
@@ -24,19 +27,11 @@ public class CreatePdfResult {
             () ->
                 runtimeEx(
                     () -> {
-                      PDDocument document = PDDocument.load(new ByteArrayInputStream(pdf));
+                      PDDocument document = Loader.loadPDF(pdf);
                       PDFTextStripper stripper = new PDFTextStripper();
 
                       pdfString = stripper.getText(document);
                       return pdfString;
                     }));
-  }
-
-  public String getHtml() {
-    return html;
-  }
-
-  public byte[] getPdf() {
-    return pdf;
   }
 }
