@@ -1,8 +1,8 @@
 package pl.matsuo.core.kv;
 
+import jakarta.persistence.MappedSuperclass;
 import java.util.ArrayList;
 import java.util.List;
-import javax.persistence.MappedSuperclass;
 import lombok.Getter;
 import pl.matsuo.core.model.kv.KeyValueEntity;
 
