@@ -1,15 +1,14 @@
 package pl.matsuo.core.model.organization;
 
+import jakarta.persistence.Entity;
+import jakarta.persistence.ManyToMany;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Transient;
 import java.util.HashSet;
 import java.util.Set;
-import javax.persistence.Entity;
-import javax.persistence.ManyToMany;
-import javax.persistence.ManyToOne;
-import javax.persistence.Transient;
 import lombok.Getter;
 import lombok.Setter;
 import org.hibernate.validator.constraints.NotEmpty;
-import pl.matsuo.validator.REGON;
 
 /**
  * Organization unit - describes company, division etc.
@@ -25,7 +24,8 @@ public class OrganizationUnit extends AbstractParty {
   private String shortName;
   private String code;
   @ManyToMany private Set<Person> employees = new HashSet<>();
-  @REGON private String regon;
+  private String regon;
+
   /** Parent organization unit. If null, entity describes standalone organization, like company. */
   @ManyToOne private OrganizationUnit parentOrganizationUnit;
 
