@@ -2,10 +2,10 @@ package pl.matsuo.core.web.controller;
 
 import static java.util.Arrays.asList;
 import static java.util.Collections.emptyList;
-import static org.hibernate.criterion.MatchMode.ANYWHERE;
 import static org.springframework.http.HttpStatus.NOT_FOUND;
 import static org.springframework.http.MediaType.APPLICATION_OCTET_STREAM_VALUE;
 import static org.springframework.web.bind.annotation.RequestMethod.GET;
+import static pl.matsuo.core.model.query.MatchMode.ANYWHERE;
 import static pl.matsuo.core.model.query.QueryBuilder.ilike;
 import static pl.matsuo.core.model.query.QueryBuilder.in;
 import static pl.matsuo.core.model.query.QueryBuilder.or;
